examples/common/requirements.txt
boto3>=1.26.0
google-cloud-compute>=1.12.0
google-cloud-storage>=2.10.0
google-cloud-kms>=2.15.0
oci>=2.90.0
cryptography>=41.0.0
pytest>=7.0.0
python-dotenv>=1.0.0 